UHF-7334: Restructured tests
