+++ tests/src/Functional/Plugin/Field/FieldFormatter/AccessibilitySentenceFormatterTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace Drupal\Tests\helfi_tpr\Functional;
+namespace Drupal\Tests\helfi_tpr\Functional\Plugin\Field\FieldFormatter;
 
 use Drupal\Core\Url;
 use Drupal\helfi_tpr\Entity\TprEntityBase;
@@ -10,6 +10,7 @@ use Drupal\helfi_tpr\Entity\TprEntityBase;
 /**
  * Tests accessibility sentence formatter.
  *
+ * @coversDefaultClass \Drupal\helfi_tpr\Plugin\Field\FieldFormatter\AccessibilitySentenceFormatter
  * @group helfi_tpr
  */
 class AccessibilitySentenceFormatterTest extends CustomFieldFormatterTestBase {
@@ -64,6 +65,9 @@ class AccessibilitySentenceFormatterTest extends CustomFieldFormatterTestBase {
 
   /**
    * Tests the formatter.
+   *
+   * @covers ::viewElements
+   * @covers ::groupItemsByLabel
    */
   public function testFormatter() : void {
     // Make sure field display is disabled by default.
